+++ service/src/components/signup/type.ts
@@ -1,3 +1,5 @@
+import { ChangeEvent, Dispatch, SetStateAction } from 'react';
+
 export interface ValueProps {
     email: string;
     nickname: string;
@@ -5,4 +7,18 @@ export interface ValueProps {
 
 export interface IParentProps {
     onSignupNext: () => void;
+}
+
+export interface ISingupFormProps {
+    onSignupNext: () => void;
+    handleClose: () => void;
+}
+
+export interface IEmailProps {
+    onEmailCertificationButton: () => void;
+    onEmailAuthenticationChange: (e: ChangeEvent<HTMLInputElement>) => void;
+    onInputChange: (e: ChangeEvent<HTMLInputElement>) => void;
+    onCount: number;
+
+    onSetCount: Dispatch<SetStateAction<number>>;
 }
