+++ service/src/components/social/SocialCallback.tsx
@@ -6,20 +6,26 @@ import { SocialService, SignInService } from '../../service/auth/SocialService';
 import { modalOpenState } from '../../recoil/atom/modalOpenState';
 import { providerIdState } from '../../recoil/atom/providerIdState';
 import { setAuthStorage } from '../../repository/AuthRepository';
+import { profileImageUrlState } from '../../recoil/atom/profileImageUrlState';
 
 export default function SocialCallback() {
     const [didMount, setDidMount] = useState(false);
 
-    const setProviderState = useSetRecoilState(providerIdState);
+    const setProviderIdState = useSetRecoilState(providerIdState);
     const code = new URL(document.location.toString()).searchParams.get('code');
     const setModalOpen = useSetRecoilState(modalOpenState);
+    const setImgUrl = useSetRecoilState(profileImageUrlState);
     const location = useLocation();
     const state = location.pathname.split('/')[1];
     const ACCESSTOKEN_KEY = 'accessToken';
     const REFRESHTOKEN_KEY = 'refreshToken';
     const navigate = useNavigate();
 
-    async function socialLoginRender(isRegistered: string, providerId: string) {
+    async function socialLoginRender(
+        isRegistered: string,
+        providerId: string,
+        profileImageUrl: string,
+    ) {
         if (isRegistered) {
             const authData = await SignInService(providerId);
             setAuthStorage(
@@ -28,18 +34,19 @@ export default function SocialCallback() {
                 REFRESHTOKEN_KEY,
                 authData.refreshToken,
             );
-            setProviderState(providerId);
+            setProviderIdState(providerId);
             navigate('/');
         } else {
             navigate('/');
-            setProviderState(providerId);
+            setProviderIdState(providerId);
+            setImgUrl(profileImageUrl);
             setModalOpen(true);
         }
     }
 
     const handleKakaoLogin = async () => {
         const data = await SocialService(code, state);
-        socialLoginRender(data.isRegistered, data.providerId);
+        socialLoginRender(data.isRegistered, data.providerId, data.profileImageUrl);
     };
 
     useEffect(() => {
