+++ service/src/recoil/atom/profileImageUrlState.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const profileImageUrlState = atom({
+    key: 'profileImageUrlState',
+    default: '',
+});
